requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Pydantic for data validation
pydantic==2.5.0

# Vector database and similarity search
pinecone-client==2.2.4
sentence-transformers==2.2.2

# OpenAI API
openai==1.3.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Environment variables
python-dotenv==1.0.0

# HTTP requests
httpx==0.25.2

# Async support
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Logging and monitoring
structlog==23.2.0

# Additional utilities
numpy==1.24.4
torch==2.1.1
transformers==4.35.2

# Document processing
PyMuPDF==1.23.8
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.1.4
nltk==3.8.1
python-multipart==0.0.6

# Development dependencies (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Security
cryptography==41.0.8
